leftclickcast: enable casting on npc, remove bloat, add more hotkeys.
